avcodec/dca_xll: fix for some DTS:X streams being detected as DTS:X IMAX
Based on the investigation and testing performed by MediaInfo team and other contributors
it was discovered that one of the two DTS:X IMAX syncwords in most cases indicates a 'standard' DTS:X stream.
This commit updates the DTS:X recognition logic to be in line with the aforementioned findings.
Despite this change there are still some false-positives due to detection being
based on the reverse engineering and not on the official specification (which is not available to mere mortals).

Related MediaInfo issue tickets:
https://github.com/MediaArea/MediaInfo/issues/717
https://github.com/MediaArea/MediaInfo/issues/823